Renames the Vector2D.hpp and Vector3D.hpp files to Vector2.hpp and Vector3.hpp respectively.
