+++ PlayRho/Common/Math.hpp
@@ -24,8 +24,8 @@
 #include <PlayRho/Common/BoundedValue.hpp>
 #include <PlayRho/Common/Span.hpp>
 #include <PlayRho/Common/UnitVec2.hpp>
-#include <PlayRho/Common/Vector2D.hpp>
-#include <PlayRho/Common/Vector3D.hpp>
+#include <PlayRho/Common/Vector2.hpp>
+#include <PlayRho/Common/Vector3.hpp>
 #include <PlayRho/Common/Position.hpp>
 #include <PlayRho/Common/Velocity.hpp>
 #include <PlayRho/Common/Acceleration.hpp>
