+++ PlayRho/Common/Sweep.hpp
@@ -24,7 +24,7 @@
 
 #include <PlayRho/Common/Position.hpp>
 #include <PlayRho/Common/Settings.hpp>
-#include <PlayRho/Common/Vector2D.hpp>
+#include <PlayRho/Common/Vector2.hpp>
 
 namespace playrho
 {
